Add two factor login support
Degiro allows two factor authentication with Google
Authenticator. When you enable this for your profile,
the default login page redirects you to a new page.
There you have to enter the code that is generated by the
Google Authenticator app.

Despite being a new page, the login POST request is
identical to the normal login, with one extra parameter.

I added this login option to the code. The code checks if
the oneTimePassword contains a value. If that is the case,
the two factor authentication login is used. Otherwise, the
default login is triggered.